Move error_handler.py to _internal
