+++ certbot/_internal/client.py
@@ -24,7 +24,7 @@ from certbot import cli
 from certbot import constants
 from certbot import crypto_util
 from certbot import eff
-from certbot import error_handler
+from certbot._internal import error_handler
 from certbot import errors
 from certbot import interfaces
 from certbot import storage
