+++ certbot/tests/error_handler_test.py
@@ -1,4 +1,4 @@
-"""Tests for certbot.error_handler."""
+"""Tests for certbot._internal.error_handler."""
 import contextlib
 import signal
 import sys
@@ -40,10 +40,10 @@ def send_signal(signum):
 
 
 class ErrorHandlerTest(unittest.TestCase):
-    """Tests for certbot.error_handler.ErrorHandler."""
+    """Tests for certbot._internal.error_handler.ErrorHandler."""
 
     def setUp(self):
-        from certbot import error_handler
+        from certbot._internal import error_handler
 
         self.init_func = mock.MagicMock()
         self.init_args = set((42,))
@@ -132,10 +132,10 @@ class ErrorHandlerTest(unittest.TestCase):
 
 
 class ExitHandlerTest(ErrorHandlerTest):
-    """Tests for certbot.error_handler.ExitHandler."""
+    """Tests for certbot._internal.error_handler.ExitHandler."""
 
     def setUp(self):
-        from certbot import error_handler
+        from certbot._internal import error_handler
         super(ExitHandlerTest, self).setUp()
         self.handler = error_handler.ExitHandler(self.init_func,
                                                  *self.init_args,
