--- docs/api/error_handler.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.error_handler`
---------------------------------
-
-.. automodule:: certbot.error_handler
-   :members:
